Fix 64-bit Linux build and MinGW cross-compilation with 64-bit Linux

+++ engine/source/gamelib/soundmix.c
@@ -364,7 +364,7 @@ int sound_load_sample(char *filename, char *packfilename, int iLog)
     lc(convcache, strlen(convcache));
     if(List_FindByName(&samplelist, convcache))
     {
-        cache = &soundcache[(int)List_Retrieve(&samplelist)];
+        cache = &soundcache[(size_t)List_Retrieve(&samplelist)];
         if(!cache->sample.sampleptr)
         {
             if(!sound_reload_sample(cache->index) && iLog)
@@ -389,7 +389,7 @@ int sound_load_sample(char *filename, char *packfilename, int iLog)
     soundcache[sound_cached].sample = sample;
 
     List_GotoLast(&samplelist);
-    List_InsertAfter(&samplelist, (void *)sound_cached, convcache);
+    List_InsertAfter(&samplelist, (void *)(size_t)sound_cached, convcache);
     soundcache[sound_cached].index = sound_cached;
     soundcache[sound_cached].filename = List_GetName(&samplelist);
